refactor: stepper에서 토글정보관리 훅으로 분리
